Agregando funciones de index, createUser, write al users.model
