+++ HC_Webapp/src/model/users.model.js
@@ -1,6 +1,7 @@
-/* const { readFileSync, writeFileSync } = require('fs');
+const { readFileSync, writeFileSync } = require('fs');
 const { resolve } = require('path');
 const { hashSync } = require('bcryptjs');
+const userLoggedIn = require('../views/users/userSession_JSON');
 
 const usersModel = {
   index: function () {
@@ -8,33 +9,27 @@ const usersModel = {
     let usersJSON = readFileSync(usersFile);
     return JSON.parse(usersJSON);
   },
-  one: function (id) {
-    let file = resolve(__dirname, '../data', 'users.json');
-    let data = readFileSync(file);
-    let users = JSON.parse(data);
-    return users.find(user => user.id === id);
-  },
-  create: function (data) {
-    let file = resolve(__dirname, '../data', 'users.json');
-    let info = readFileSync(file);
-    let users = JSON.parse(info);
-    let last = users[users.length - 1];
-    return Object({
-      id: users.length == 0 ? 1 : last.id + 1,
-      nombre: data.nombre,
-      apellido: data.apellido,
-      email: data.email,
-      password: hashSync(data.password, 10),
-      image: data.image,
-      isAdmin: data.email.includes('@dh.com'),
-    });
+  createUser: function (data) {
+    let usersList = usersModel.index();
+    let tempID = usersList.length;
+    tempID++;
+    let tempUser = {
+      id: tempID,
+      first_name: data.first_name,
+      last_name: data.last_name,
+      user_name: data.user_name,
+      passwd: hashSync(data.passwd, 10),
+      role: '',
+      loggedIn: false,
+    };
+    usersList.push(tempUser);
+    usersModel.write(usersList);
   },
   write: function (data) {
-    let file = resolve(__dirname, '../data', 'users.json');
-    let info = JSON.stringify(data, null, 2);
-    return writeFileSync(file, info);
+    let usersFile = resolve(__dirname, '../data', 'usersList.json');
+    let update = JSON.stringify(data, null, 2);
+    return writeFileSync(usersFile, update);
   },
 };
 
 module.exports = usersModel;
- */
