+++ HC_Webapp/src/routes/login.routes.js
@@ -2,6 +2,7 @@ const express = require('express');
 const router = express.Router();
 const userLoggedIn = require('../views/users/userSession_JSON');
 const controllerLogin = require('../controllers/controller.login.js');
+const { index, createUser, write } = require('../model/users.model');
 const { resolve } = require('path');
 const { readFileSync, writeFileSync } = require('fs');
 
@@ -10,9 +11,7 @@ router.get('/', controllerLogin.mostrarLogin);
 router.get('/close', controllerLogin.closeSession);
 
 router.put('/update', (req, res) => {
-  let usersFile = resolve(__dirname, '../data', 'usersList.json');
-  let usersJSON = readFileSync(usersFile);
-  let usersList = JSON.parse(usersJSON);
+  let usersList = index();
   usersList.forEach(function (user, index) {
     if (userLoggedIn.user_name == user.user_name) {
       this[index].first_name = req.body.first_name;
@@ -25,18 +24,14 @@ router.put('/update', (req, res) => {
       userLoggedIn.passwd = req.body.passwd;
     }
   }, usersList);
-  let write = JSON.stringify(usersList, null, 2);
-  writeFileSync(usersFile, write);
+  write(usersList);
   res.redirect('/');
 });
 
 router.get('/user', controllerLogin.accountDetails);
 
 router.put('/user', (req, res) => {
-  let validateUser = 'el usuario no existe o la contraseña es incorrecta';
-  let usersFile = resolve(__dirname, '../data', 'usersList.json');
-  let usersJSON = readFileSync(usersFile);
-  let usersList = JSON.parse(usersJSON);
+  let usersList = index();
   usersList.forEach(user => {
     if (
       req.body.user_name == user.user_name &&
@@ -47,7 +42,6 @@ router.put('/user', (req, res) => {
       userLoggedIn.last_name = user.last_name;
       userLoggedIn.user_name = user.user_name;
       userLoggedIn.passwd = user.passwd;
-      validateUser = user;
     }
   });
   res.redirect('/');
