+++ HC_Webapp/src/routes/register.routes.js
@@ -4,27 +4,12 @@ const controllerRegister = require('../controllers/controller.register.js');
 const userLoggedIn = require('../views/users/userSession_JSON');
 const { resolve } = require('path');
 const { readFileSync, writeFileSync } = require('fs');
+const { index, createUser, write } = require('../model/users.model');
 
 router.get('/', controllerRegister.mostrarRegister);
 
 router.put('/user', (req, res) => {
-  let usersFile = resolve(__dirname, '../data', 'usersList.json');
-  let usersJSON = readFileSync(usersFile);
-  let usersList = JSON.parse(usersJSON);
-  let tempID = usersList.length;
-  tempID++;
-  let tempUser = {
-    id: tempID,
-    first_name: req.body.first_name,
-    last_name: req.body.last_name,
-    user_name: req.body.user_name,
-    passwd: req.body.passwd,
-    role: '',
-    loggedIn: false,
-  };
-  usersList.push(tempUser);
-  let write = JSON.stringify(usersList, null, 2);
-  writeFileSync(usersFile, write);
+  createUser(req.body);
   res.redirect('/');
 });
