--- package.json
@@ -1,24 +1,0 @@
-{
-  "name": "handicraft",
-  "version": "2.0.0",
-  "main": "app.js",
-  "scripts": {
-    "test": "node app.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "ejs": "^3.1.8",
-    "express": "^4.18.1",
-    "express-session": "^1.17.3",
-    "method-override": "^3.0.0",
-    "multer": "^1.4.5-lts.1",
-    "normalize.css": "^8.0.1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/fcgrandoli"
-  },
-  "description": "Es un emprendimiento que pone al alcance de un click la compra de productos en los que interviene el diseño de articulos en vidrio, con los mejores estándares de especificación y calidad."
-}
